Fix for large ASCII STL files causing OOM
Large ASCII STL files were causing OOM errors in parsing, due to
the way STLLoader was converting from binary to string. On a 73 MB
file this created too many intermediate strings for GC to handle.

Switches to the same version of STLLoader we use in the client side,
which includes a check for the browser-based TextDecoder interface
to do this string conversion much more efficiently...

...so we add a quick polyfill for TextDecoder to use Node's Buffer
API to do the same thing. This polyfill strategy is recommended by
the three.js authors.

Bug: T188611
Change-Id: Ifbf476af3df54bebdc7910850d93d75e8819f22d